% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{itemModelExplorer}
\alias{itemModelExplorer}
\title{A Shiny app to experiment with the item response model}
\usage{
itemModelExplorer(dl = NULL, fit = NULL, item = NULL)
}
\arguments{
\item{dl}{a data list prepared by \code{\link{prepData}}}

\item{fit}{a \code{\link[rstan:stanfit-class]{stanfit}} object}

\item{item}{name of the item to visualize}
}
\description{
When data \code{dl} and fitted model \code{fit} are provided, the
item parameters associated with \code{item} are loaded for
inspection.
}
\section{Response model}{


See \link{cmp_probs} for details.
}

\examples{
\donttest{
itemModelExplorer()  # will launch a browser in RStudio
}
}
