% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{roundRobinGraph}
\alias{roundRobinGraph}
\title{Create an edge list with round-robin connectivity}
\usage{
roundRobinGraph(name, N)
}
\arguments{
\item{name}{vector of vertex names}

\item{N}{number of comparisons}
}
\value{
An undirected graph represented as a data frame with each row describing an edge.
}
\description{
Create an edge list with round-robin connectivity
}
\examples{
roundRobinGraph(letters[1:5], 10)
}
\seealso{
Other graph generators: \code{\link{twoLevelGraph}}
}
\concept{graph generators}
