% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{unfactor}
\alias{unfactor}
\title{Turn a factor back into a vector of integers}
\usage{
unfactor(f)
}
\arguments{
\item{f}{a factor}
}
\description{
Factors store values as integers and use a 'levels' attribute to
map the integers to labels. This function removes the 'factor'
class and levels attribute, leaving the vector of integers.
}
\examples{
f <- factor(letters[1:3])
print(f)
print(unfactor(f))
}
