% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{withoutIndex}
\alias{withoutIndex}
\title{Remove the array indexing from a parameter name}
\usage{
withoutIndex(name)
}
\arguments{
\item{name}{a parameter name}
}
\value{
the name without the square bracket parameter indexing
}
\description{
Remove the array indexing from a parameter name
}
\examples{
withoutIndex("foo[1,2]")
}
