% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{itemModelExplorer}
\alias{itemModelExplorer}
\title{A Shiny app to experiment with the item response model}
\usage{
itemModelExplorer()
}
\description{
A Shiny app to experiment with the item response model
}
\examples{
\donttest{
itemModelExplorer()  # will launch a browser in RStudio
}
}
